# documentacion/DOCUMENTACION_PAWMILY/03_Manuales_Tecnicos_y_DevOps/Especificacion_APIs_OpenAPI.yaml
openapi: 3.0.3
info:
  title: PawMily API
  version: "1.0.0"
  description: |
    API REST de PawMily (Express Clean Architecture + DDD).
    Base producción: https://api-production-66b1.up.railway.app/api
  contact:
    name: Equipo PawMily
servers:
  - url: https://api-production-66b1.up.railway.app/api
    description: Producción Railway
  - url: http://127.0.0.1:3000/api
    description: Desarrollo local
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Registro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                name: { type: string }
                role: { type: string, enum: [vet, owner] }
      responses:
        "201": { description: Usuario creado / tokens }
        "400": { description: Validación }
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: accessToken + refreshToken + user
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Renovar access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200": { description: Nuevo accessToken }
...
  /auth/profile:
    get:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      summary: Perfil actual
      responses:
        "200": { description: User }
    put:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      summary: Actualizar perfil
      responses:
        "200": { description: User actualizado }
...
  /appointments:
    get:
      tags: [Appointments]
      security: [{ bearerAuth: [] }]
      summary: Listar citas (vet)
      responses:
        "200": { description: data }
    post:
      tags: [Appointments]
      security: [{ bearerAuth: [] }]
      summary: Crear cita
      responses:
        "201": { description: Appointment }
  /appointments/mine:
    get:
      tags: [Appointments]
      security: [{ bearerAuth: [] }]
      summary: Citas del owner
      responses:
        "200": { description: data }
  /appointments/month:
    get:
      tags: [Appointments]
      security: [{ bearerAuth: [] }]
      summary: Citas por mes
      responses:
        "200": { description: data }
  /appointments/{id}/confirm:
    post:
      tags: [Appointments]
      security: [{ bearerAuth: [] }]
      summary: Confirmar asistencia (owner)
      responses:
        "200": { description: Appointment }
...
  /favorites:
    get:
      tags: [Favorites]
      security: [{ bearerAuth: [] }]
      summary: Listar favoritos
      responses:
        "200": { description: Lista }
    post:
      tags: [Favorites]
      security: [{ bearerAuth: [] }]
      summary: Agregar favorito
      responses:
        "201": { description: Favorite }
  /media/upload-url:
    post:
      tags: [Media]
      security: [{ bearerAuth: [] }]
      summary: URL firmada de subida
      responses:
        "200": { description: upload URL }
  /media/{assetId}/url:
    get:
      tags: [Media]
      security: [{ bearerAuth: [] }]
      summary: URL de lectura de asset
      parameters:
        - in: path
          name: assetId
          required: true
          schema: { type: string }
      responses:
        "200": { description: url }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...